chore: use Biome for linting and formatting
[Biome][1] is a next-generation [opinionated][2] formatter and linter.
It is extremely fast and by using one tool for both formatting and
linting we avoid the conflicts that can often arise when using
Prettier and ESLint.

This commit installs Biome as per the [installation docs][3].
It also adds [the VS Code Biome extension][4] to our devcontainer
(Codespace), and sets it up to format files automatically on save.
Not using Biome to organise imports (yet), as it seemed buggy when we
tried it.

[1]: https://biomejs.dev/
[2]: https://biomejs.dev/formatter/option-philosophy/
[3]: https://biomejs.dev/guides/getting-started/#installation
[4]: https://biomejs.dev/reference/vscode/