chore: remove ESLint pre-commit staged lint check
We don't need this check anymore now that we are using Biome.  We could
invoke Biome as a pre-commit check, but don't think that is necessary as
our convention should be to fix any Biome linting issues before we
commit, with the safeguard of the Biome CI check to make sure we're
always ok.