chore: add PostgreSQL to Codespace
Adding PostgreSQL to our Codespace led us to modify our definition to
use two containers rather than one.  In addition to the Node.js with
TypeScript container which we were already using, we now also have a
PostgreSQL DB container. Having two containers means that we now also
use [Docker Compose][1] to manage them.

The documentation for how to use the PostgreSQL database is in the
README in the `.devcontainer` directory.

[1]: https://docs.docker.com/compose/