chore: ensure that only maintainers approve PRs (#25)
We only want project maintainers to be able to approve PRs.  For
instance we don't want the `coliving-semkovo` GitHub account to be able
to approve PRs.  We achieve this via the [`CODEOWNERS`][1] file,
together with enabling the [protected branch setting][2] which requires
each PR to be reviewed and approved by a `CODEOWNER`.

[1]: https://help.github.com/articles/about-codeowners/
[2]: https://help.github.com/articles/about-protected-branches/