chore: exclude package-lock.json from Biome
Biome treats package-lock.json as a protected file and so doesn't
format it (though prior to updating Biome to 1.6.x there was a bug where
the Biome VS Code extension would still format it).  Even though Biome
will not format the package-lock.json now, we still explicitly list it
as an ignored file in the biome.json, as otherwise VS Code goes through
the rigmarole of asking Biome to format it, and then reporting that
Biome does not format it.