chore: check env vars at build time (#27)
* chore: check env vars at build time

We were already following [12 factor principles][1] by
[storing our config in the environment][2].  Because we store our config
in the environment it is important to validate that there are no missing
environment variables, and also ideally to validate that they are in the
expected format (e.g. a database URL is properly formatted) where
appropriate.

Prior to this commit we were not validating our environment config. We
now check that all required environment variables are present.  We
validate this at build time as we want to fail fast if there are any
problems, rather than deploying a misconfigured app to our customers
(which would happen if we only validated our config at runtime, or
didn't validate at all).

We followed [the initial step in this excellent article][3] to achieve
this. NB the article suggests doing some fairly fancy stuff to check
required client env vars, not just server-side, but that seems unnecessary
as we can just check on the server-side for any required `NEXT_PUBLIC_`
env vars, and rely on Next.js to expose them client-side.

[1]: https://12factor.net
[2]: https://12factor.net/config
[3]: https://sophiabits.com/blog/verify-environment-variables-nextjs

* fix: expose required env vars in GitHub Actions

* chore: error if server check imported on frontend

Added this as per the suggestion in
[this article][1].

[1]: https://sophiabits.com/blog/verify-environment-variables-nextjs

* chore: check database env var format

This is an improvement over just checking if the env var is present,
as now we will fail fast (i.e. at build time) if the `DATABASE_URL`
is not in the correct format for a PostgreSQL connection URL.  Added
as per the suggestion in [this article][1].

[1]: https://sophiabits.com/blog/verify-environment-variables-nextjs