chore: add checkjs error reporting for js files
Having error reporting for js files is a good thing, so good to enable
this via [checkjs][1].
The `env.mjs` file failed the error checking, complaining that the
vars prefixed with `NEXT_PUBLIC_` should not have this prefix, for some
reason. So we exclude those lines from the error checking for now, and
will investigate and hopefully fix them at a later point.

[1]: https://www.typescriptlang.org/tsconfig#checkJs