chore: use correct is production logic for Prisma
The example in the Prisma docs deems production to be when `NODE_ENV` is
"production", but that is not what we need, as for instance the
Vercel preview environments would be deemed production. We need our
logic here to check for the real live production environment only, so we
can inject a new Prisma client with a unique DB schema when running
Playwright tests on any other environment (including Vercel preview
environments).